requirements.txt
numpy==2.4.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
pillow==12.1.1
Jinja2==3.1.6
MarkupSafe==3.0.3
filelock==3.21.2
fsspec==2026.2.0
networkx==3.6.1
mpmath==1.3.0
sympy==1.14.0
typing_extensions==4.15.0
